refactor: organizando codigos e pastas
