--- projeto-engenharia-software/src/main/java/CadastraUsuario/CadastraUsuario.java
@@ -1,33 +1,0 @@
-package CadastraUsuario;
-
-public class CadastraUsuario {
-    private String nome;
-    private String email;
-    private String senha;
-    private String nivel_permissao;
-    
-    public String getNome() {
-        return nome;
-    }
-    public void setNome(String nome) {
-        this.nome = nome;
-    }
-    public String getEmail() {
-        return email;
-    }
-    public void setEmail(String email) {
-        this.email = email;
-    }
-    public String getSenha() {
-        return senha;
-    }
-    public void setSenha(String senha) {
-        this.senha = senha;
-    }
-    public String getNivel_permissao() {
-        return nivel_permissao;
-    }
-    public void setNivel_permissao(String nivel_permissao) {
-        this.nivel_permissao = nivel_permissao;
-    }  
-}
