+++ projeto-engenharia-software/src/main/java/dao/CadastraUsuarioDAO.java
@@ -3,14 +3,13 @@ package dao;
 import java.sql.Connection;
 import java.sql.PreparedStatement;
 import java.sql.SQLException;
-
-import CadastraUsuario.CadastraUsuario;
+import dominio.Usuario;
 import dominio.Conexao;
 
 public class CadastraUsuarioDAO {
     private static final String  Insert_SQL = "INSERT INTO usuario (nome, email, senha, nivel_permissao) values (?,?,?,?)";
     
-    public static boolean cadastraUsuario(CadastraUsuario u){
+    public static boolean cadastraUsuario(Usuario u){
 		boolean sucesso = false;
         
         try {
